// src/app/components/interfaces/orders.interfaces.ts
export interface infoClient{
  city: String;
  plant: String;
}
export interface infoPlant{
  name_client:string;
  city: String;
  plant: String[];
  plant2:String;
}
export interface infoPart {
  plant:        string;
  no_parte:     string;
  descrip:      string;
  no_orden:     string;
  clave:        String;
  medida:       string;
  pieza_medida: number;
  articulo:     number;
  pieza_caja:   number;
}

export interface infoInputs {
  plant:        string;
  no_parte:     string;
  descrip:      string;
  no_orden:     string;
  clave:        String;
  medida:       string;
  pieza_medida: number;
  articulo:     number;
  pieza_caja:   number;
  num_visual:   string;
  autorizo:     string;
  stock:number;

}
// ...
export interface stock {
  stock:number;

}